refactor(evals): extract shared werewolf rules (roles, action parsing, win) into werewolf-rules.ts
Pure pieces only, behavior identical — the Slack-shaped game delegates to the
shared module so the upcoming webchat composition cannot drift on the rules.
